refactor: reorganize usecase package and implement GeminiAI and LineBotClient functionality
