+++ api/usecase/gemini_ai.go
@@ -1,4 +1,4 @@
-package service
+package usecase
 
 import (
 	"context"
