+++ api/usecase/linebot.go
@@ -1,4 +1,4 @@
-package service
+package usecase
 
 import (
 	"fmt"
