+++ api/usecase/recipe_service.go
@@ -1,4 +1,4 @@
-package service
+package usecase
 
 import (
 	"context"
